+++ tests/model.rs
@@ -20,7 +20,7 @@ impl Arbitrary for Op {
     let offset: u64 = g.gen_range(0, MAX_FILE_SIZE);
     let length: u64 = g.gen_range(0, MAX_FILE_SIZE / 3);
 
-    let op = g.gen_range(0 as u8, 3 as u8);
+    let op = g.gen_range(0_u8, 3_u8);
 
     if op == 0 {
       Read { offset, length }
